+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - Added test to enforce `CHANGELOG.md` updates in PRs to `dev` (`.github/workflows/check-changelog.yml`)
     - Consolidated version & output tracking into `pyproject.toml` & added a test for version consistency between `pyproject.toml` and `CHANGELOG.md`.
     - Moved release documentation from private internal docs to `docs/developer.md` and updated formatting to match Github requirements.
+    - Added verification of expected outputs to workflow tests.
 
 # v3.0.1.7
 - Clarified testing documentation in `docs/developer.md`.